+++ applications/lobsters/src/queriers/comment.rs
@@ -20,7 +20,7 @@ pub fn get_comments(db: &mut mysql::PooledConn, acting_as: Option<u64>) -> Resul
              AND `comments`.`is_moderated` = 0 \
              ORDER BY id DESC \
              LIMIT 40",// OFFSET 0",
-        ), |(id, user_id, story_id) : (u32, u32, u32)| {
+        ), |(id, user_id, story_id) : (u32, u64, u32)| {
             comments.insert(id);
             users.insert(user_id);
             stories.insert(story_id);
@@ -66,7 +66,7 @@ pub fn get_comments(db: &mut mysql::PooledConn, acting_as: Option<u64>) -> Resul
              WHERE `stories`.`id` IN ({})",
             stories
         ),
-        |user_id: u32| authors.insert(user_id))?;
+        |user_id: u64| authors.insert(user_id))?;
     warn!("\t select authors {}", start.elapsed().as_micros());
 
     assert!(!authors.is_empty());
